requirements.txt
anyio==4.1.0
asgiref==3.7.2
certifi==2023.11.17
Django==4.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
Pillow==10.1.0
psycopg2-binary==2.9.9
python-telegram-bot==20.7
pytz==2023.3.post1
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
